check if user is authorize to edit or delete post

+++ controllers/feedController.js
@@ -143,6 +143,11 @@ module.exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if(post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
       if(imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -171,7 +176,11 @@ module.exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
-
+      if(post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
       // Check logged in user
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
